requirements.txt
flask
flask_sqlalchemy
sqlalchemy
discord.py
requests
pyqt5
pymem
gunicorn
psycopg2-binary
